Add extra tests for pytest
